fastapi/requirements.txt
numpy==1.18.1
pandas==0.25.3
xgboost==1.1.1
SQLAlchemy
holidays
redis
python-redis-cache
firebase-admin
mysqlclient